% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{beargrowth}
\alias{beargrowth}
\title{beargrowth}
\format{
A data frame with 835 rows and 4 variables:
\describe{
\item{\code{sex}}{character sex of the individual}
\item{\code{age}}{integer age of the individual in years}
\item{\code{wtkg}}{double weight in kilograms}
}
#@source
}
\usage{
beargrowth
}
\description{
Size and weigth measurements for 390 black bears monitored in Minnesota by the MN Department of Natural Resources.
}
\keyword{datasets}
