% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bearmove}
\alias{bearmove}
\title{bearmove}
\format{
A data frame with 2768 rows and 10 variables:
\describe{
\item{\code{log.move}}{double log movement rate, where movement rate is calculated using the distance between successive telemetry obsevartions divided by the time between observations.}
\item{\code{hr}}{integer heart rate measured using implanted transmitters}
\item{\code{Season}}{character Season of the observation (Fall, Spring, or Summer)}
\item{\code{DayNight}}{character Timing of the observation (Day or Night)}
\item{\code{Sex}}{character sex of the bear (Male or Female)}
\item{\code{Stage}}{character categorical variable equal to Male, Female without cubs (Fem)), or Female with Cubs (CubPrep) Pre}
\item{\code{julian}}{integer Julian day of the observation}
\item{\code{BearID}}{integer unique identifier associated with each bear}
\item{\code{Year}}{integer Year of the observation}
\item{\code{BearIDYear}}{integer unique label assigned to each combination of bear + year}
}
}
\source{
Ditmer, M. A. (2014). American black bears: strategies for living in a fragmented, agricultural landscape. PhD Dissertation: University of Minnesota.
}
\usage{
bearmove
}
\description{
data set containing observed movement rates and associated average heart rates of 7 black bears in Minnesota.
}
\keyword{datasets}
