% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{beedat}
\alias{beedat}
\title{beedat}
\format{
A data frame with 1040 rows and 6 variables:
\describe{
\item{\code{Bee}}{integer ID of the individual bee}
\item{\code{Trial}}{integer Choice number ranging from 1-20 for each test}
\item{\code{Choice}}{integer Individual choices made by each bee. Choice = 1 if a correct choice was made and 0 otherwise.}
\item{\code{Test}}{integer The learning test for even trained bees is marked as ‘1’, the learning test for odd trained bees is marked as ‘2’, the transfer test for even trained bees is marked as ‘3’, and the transfer test for odd trained bees is marked ‘4’.}
\item{\code{Ieven}}{double Indicator variable equal to 1 for even trained bees and 0 for odd trained bees}
\item{\code{Itransfer}}{double Indicator variable equal to 1 if the observation was from a learning test and 0 for a transfer test}
}
}
\source{
Howard, S. R., Greentree, J., Avarguès-Weber, A., Garcia, J. E., Greentree, A. D., & Dyer, A. G. (2022). Numerosity Categorization by Parity in an Insect and Simple Neural Network. Frontiers in Ecology and Evolution, 252.

Data were downloaded from supplemental material associated with the above paper.
}
\usage{
beedat
}
\description{
Data set containing choices made by bees during experiments used to evaluate whether they could learn to select for odd or even numbers and transfer that learning to new situations.  Data are from Howard et al. (2022)
}
\keyword{datasets}
