% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{birdmalariaLFS}
\alias{birdmalariaLFS}
\title{birdmalariaLFS}
\format{
A data frame with 100 rows and 6 variables:
\describe{
\item{\code{Ind}}{character individual bird identifier}
\item{\code{LFS}}{double lifetime reproductive success (number of offspring produced) }
\item{\code{Sex}}{double sex of the individual, 0 if male, 1 if female}
\item{\code{year}}{double year of birth}
\item{\code{infected}}{double infection status = 0 if uninfected, 1 if infected}
\item{\code{Tlifespan}}{double total lifespan (in years)}
\item{\code{BTL}}{double telomere length in birth year}
}
Data are from:
}
\source{
Asghar, M., Hasselquist, D., Hansson, B., Zehtindjiev, P., Westerdahl, H., & Bensch, S. (2015). Hidden costs of infection: chronic malaria accelerates telomere degradation and senescence in wild birds. Science, 347(6220), 436-438.

Orinal (pre-processed data) can be downloaded from: https://datadryad.org/stash/dataset/doi:10.5061/dryad.d04h0
}
\usage{
birdmalariaLFS
}
\description{
Data set containing lifetime reproductive success, sex, year of birth, and lifespan for great reed warblers from Asghar et al. 2015. Science 347:436-438.
}
\keyword{datasets}
