% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{birdmalariaO}
\alias{birdmalariaO}
\title{birdmalariaO}
\format{
A data frame with 207 rows and 10 variables:
\describe{
\item{\code{ID}}{double  individual bird (offspring) identity at birth}
\item{\code{brood}}{double brood identifier}
\item{\code{broodyear}}{double year of birth}
\item{\code{Sex}}{double Sex of the offspring (0 if male, 1 if female)}
\item{\code{OffBTL}}{double  Offspring ealy-life telomere length (at 9 day age)}
\item{\code{mmal}}{double Mother malaria status at the time of breeding (1 if infected and 0 otherwise)}
\item{\code{mage}}{double Mother age at the time of breeding}
\item{\code{dam}}{character Unique identifier for each mother}
\item{\code{Sire}}{character Unique identifier for each father}
\item{\code{fmal}}{double Father malaria status at time of breeding (1 if infected and 0 otherwise)}
\item{\code{fage}}{double Father's age at time of breeding}
}
Asghar, M., Hasselquist, D., Hansson, B., Zehtindjiev, P., Westerdahl, H., & Bensch, S. (2015). Hidden costs of infection: chronic malaria accelerates telomere degradation and senescence in wild birds. Science, 347(6220), 436-438.

Orinal (pre-processed data) can be downloaded from: https://datadryad.org/stash/dataset/doi:10.5061/dryad.d04h0
}
\usage{
birdmalariaO
}
\description{
Great reed warbler observations of offspring telomere length and mothers infection status (infected with malaria or not) from Asghar et al. 2015. Science 347:436-438.
}
\keyword{datasets}
