% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{birdmalariaTL}
\alias{birdmalariaTL}
\title{birdmalariaTL}
\format{
A data frame with 174 rows and 5 variables:
\describe{
\item{\code{Ind}}{character individual bird identifier}
\item{\code{Sex}}{double sex of the individual, 0 if male, 1 if female}
\item{\code{year}}{double  year of birth}
\item{\code{Age}}{double age of the bird}
\item{\code{infected}}{infection status = 0 if uninfected, 1 if infected}
\item{\code{TL}}{double telomere length}
}

Data are from:
}
\source{
Asghar, M., Hasselquist, D., Hansson, B., Zehtindjiev, P., Westerdahl, H., & Bensch, S. (2015). Hidden costs of infection: chronic malaria accelerates telomere degradation and senescence in wild birds. Science, 347(6220), 436-438.

Orinal (pre-processed data) can be downloaded from: https://datadryad.org/stash/dataset/doi:10.5061/dryad.d04h0
}
\usage{
birdmalariaTL
}
\description{
Observations of telomere length as well as demographics and infection status of great reed warblers from Asghar et al. 2015. Science 347:436-438.
}
\keyword{datasets}
