% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CarnivoreScat}
\alias{CarnivoreScat}
\title{CarnivoreScat}
\format{
A data frame with 68 rows and 4 variables:
\describe{
\item{\code{Predator}}{character Predator consuming prey}
\item{\code{Prey.number}}{character Type of prey and number of prey consumed in feeding trial}
\item{\code{Mean_prey_wt}}{double Mean wt of prey offered during feeding trial (kg)}
\item{\code{Biomass.per.scat}}{double Biomass (kg) consumed per collectable scat}
}
}
\source{
Chakrabarti, S., Jhala, Y. V., Dutta, S., Qureshi, Q., Kadivar, R. F., & Rana, V. J. (2016). Adding constraints to predation through allometric relation of scats to consumption. Journal of Animal Ecology, 85(3), 660-670.
}
\usage{
CarnivoreScat
}
\description{
Measures of biomass consumption relative to collectable scat
}
\keyword{datasets}
