% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{carppassage}
\alias{carppassage}
\title{carppassage}
\format{
A data frame with 143 rows and 7 variables:
\describe{
\item{\code{Pass}}{integer equal to 1 if the fish entered the lock and 0 otherwise}
\item{\code{Approaches}}{integer nubmer of 15-min intervals the fish was detected near the lock entrance}
\item{\code{Qave}}{double daily discharge in cubic meters per second }
\item{\code{Tave}}{double average daily water temperature in Celsius}
\item{\code{Sound}}{integer sound treatment equal to 1 if sound was on and 0 otherwise}
\item{\code{FishID}}{character Tag ID code unique to each individual fish}
\item{\code{Date}}{character COLUMN_DESCRIPTION}
}
}
\source{
Data were downloaded from \url{https://conservancy.umn.edu/handle/11299/226079} and are described in
Riesgraf AT, Finger JS, Zielinski DP, Dennis III CE, Whitty JM, Sorensen PW (2022) Evaluation of a broadband sound projected from the gates of a navigation lock in the Mississippi River shows it to be a weak deterrent for common carp and unable to block passage. Management of Biological Invasions 13
}
\usage{
carppassage
}
\description{
Data from experimental trials aimed at blocking passage of carp in lock and dams along the Missippi River.
}
\keyword{datasets}
