% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clutch}
\alias{clutch}
\title{clutch}
\format{
A data frame with 139 rows and 9 variables:
\describe{
\item{\code{year}}{integer year of observation}
\item{\code{strtno}}{integer unique identifier associated with each nesting structure}
\item{\code{pairs}}{double estimated number of Mallard pairs with access to the structure}
\item{\code{agr_area}}{double cropland area within a 32.2 km^2 cirucular buffur of the nest}
\item{\code{DEPLY}}{integer structure type (equal to 1 for single-cylinder structures and 2 for double-cylinder structures)}
\item{\code{CYLDR}}{integer cylinder number (1 or 2 for double-cylinder structures)}
\item{\code{CLUTCH}}{integer clutch size}
\item{\code{prcnt_area}}{double percent of the 32.2 km^2 cirucular buffur associated with crops}
\item{\code{date}}{integer Julian date (number of days since January 1 of the current year)}
}
}
\source{
Zicus, M. C., Fieberg, J., & Rave, D. P. (2003). Does mallard clutch size vary with landscape composition: a different view. The Wilson Journal of Ornithology, 115(4), 409-413.
}
\usage{
clutch
}
\description{
Data set containing Mallard (\emph{Anas platyhynchos}) clutch sizes in two types of nest structures and cropland area within a 3.2-km radius circular buffer
of each structure in the landscape in western Minnesota during 1997–1999.
}
\keyword{datasets}
