% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EggReject}
\alias{EggReject}
\title{EggReject}
\format{
A data frame with 57 rows and 7 variables:
\describe{
\item{\code{nest.ID}}{character nest id}
\item{\code{site}}{character site either Dongzhai, Henan or  Zuojia, Jilin}
\item{\code{clutch.size}}{integer number of eggs in the clutch}
\item{\code{exp.date}}{double indicates the date when a painted blue white-rumped munia egg was added to the nest}
\item{\code{Julian.date}}{integer Julian date corresponding to exp.date}
\item{\code{Result}}{character response of female tits to the addition of the painted blue white-rumped munia egg（“rejected” if the egg was rejected and “not” if the egg was accepted)}
\item{\code{Behavioral.Type}}{character response of incubating female tits to the approaching humans (“shy” individuals left the nest whereas "bold" individuals did not)}
}
}
\source{
Shen, C., Yu, J., Yin, D., Liu, Q., Wang, H., & Liang, W. (2023). Bold–shy continuum does not account for egg rejection behaviour in the Japanese tit. Biological Journal of the Linnean Society, 140(1), 33-40.

Shen, Chao et al. (2023). Data for: Shy-boldness cannot predict egg rejection in the Japanese tit \link{Dataset}. Dryad. https://doi.org/10.5061/dryad.vt4b8gtx4

Data names modified from the original form contained on Dryad via the following link:
https://datadryad.org/stash/dataset/doi:10.5061/dryad.vt4b8gtx4
}
\usage{
EggReject
}
\description{
Experiment to evaluate differential responses to nest parasitism for bold and shy individuals
}
\keyword{datasets}
