% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elephantmove}
\alias{elephantmove}
\title{elephant move}
\format{
A data frame with 3766 rows and 9 variables:
\describe{
\item{\code{animal_id}}{integer unique identifier associated with each individual elephant}
\item{\code{Week_id}}{integer unique identifier for each week in the data set}
\item{\code{meanDailyDisplacement}}{double mean daily movement distance (km)}
\item{\code{meanRT}}{double mean residence time}
\item{\code{tac}}{double turn angle correlation}
\item{\code{week}}{double week of obsevation}
\item{\code{month}}{double month of observation}
\item{\code{year}}{double year of observation}
\item{\code{Sex}}{integer sex of the individual}
}
}
\source{
Abrahms, B., Seidel, D. P., Dougherty, E., Hazen, E. L., Bograd, S. J., Wilson, A. M., ... & Getz, W. M. (2017). Suite of simple metrics reveals common movement syndromes across vertebrate taxa. Movement ecology, 5(1), 1-11.

Tsalyuk, M., Kilian, W., Reineking, B., & Getz, W. M. (2019). Temporal variation in resource selection of African elephants follows long‐term variability in resource availability. Ecological Monographs, 89(2), e01348.

Hertel, A. G., Niemelä, P. T., Dingemanse, N. J., & Mueller, T. (2020). A guide for studying among-individual behavioral variation from movement data in the wild. Movement ecology, 8(1), 1-18.
}
\usage{
elephantmove
}
\description{
Movement data from 35 African elephants between 2008 and 2015 from Abrahms et al. (2017), Tsalyuk et al. (2019), Hertel et al. (2020).
}
\keyword{datasets}
