% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fish}
\alias{fish}
\title{fish}
\format{
A data frame with 250 rows and 6 variables:
\describe{
\item{\code{count}}{integer number of fish caught}
\item{\code{livebait}}{integer equal to 1 if live bait was used and 0 otherwise}
\item{\code{camper}}{integer equal to 1 if the group had a camper and 0 otherwise}
\item{\code{persons}}{integer the number of visitors in each group}
\item{\code{child}}{integer the number of children in each group}
}
}
\source{
\url{https://stats.idre.ucla.edu/stat/data/fish.dat} and \url{https://stats.oarc.ucla.edu/r/dae/zip/}
}
\usage{
fish
}
\description{
Data used to predict the number of fish caught by groups of visitors of a state park.
}
\keyword{datasets}
