% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Fish_Mercury}
\alias{Fish_Mercury}
\title{Fish_Mercury}
\format{
A data frame with 378 rows and 4 variables:
\describe{
\item{\code{species}}{integer species common name}
\item{\code{length}}{double length of the fish (mm)}
\item{\code{Hg}}{double mercury concentration (ug/wet g)}
\item{\code{state}}{integer State where fish was sampled}
}
}
\source{
U.S. Environmental Protection Agency. (2021). National Aquatic Resource Surveys. Rivers and Streams, Fish Tissue (Plugs) - Mercury 2018/2019 (data and metadata files). Available from U.S. EPA website: http://www.epa.gov/national-aquatic-resource-surveys/data-national-aquatic-resource-surveys. Date accessed: 2024-02-24.
}
\usage{
Fish_Mercury
}
\description{
Mercury concentrations in various fish species collected across a number of states from 2018 to 2019.
}
\keyword{datasets}
