% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fishmorph}
\alias{fishmorph}
\title{fishmorph}
\format{
A data frame with 45 rows and 13 variables:
\describe{
\item{\code{Specimen}}{character Specimen ID}
\item{\code{Species}}{character Species of fish}
\item{\code{Diet_sp}}{character Diet = C (carnivore), H (herbivore), or O (omnivore)}
\item{\code{TL}}{integer Total Length (in mm)}
\item{\code{skull_length}}{double Skull length (in mm)}
\item{\code{lj_length}}{double Lower jaw length (in mm)}
\item{\code{lj_closing}}{double Lower jaw length when closed (in mm)}
\item{\code{lj_opening}}{double Lower jaw length when open (in mm)}
\item{\code{jaw_width}}{double Jaw width (in mm)}
\item{\code{premaxilla_height}}{double Height of premaxilla  (in mm)}
\item{\code{premaxilla_length}}{double Lenght of premaxilla (in mm)}
\item{\code{maxilla_length}}{double Length of maxilla  (in mm)}
\item{\code{UJ_length}}{double Universal Joint length (in mm)}
}
}
\source{
These data were contributed by \href{https://www.bellmuseum.umn.edu/grad-awards/2024-graduate-award-and-fellowship-recipients/richard-rc-hoover/}{Richard Hoover} and \href{https://www.kassandraford.com/}{Dr. Kassandra Ford}
}
\usage{
fishmorph
}
\description{
Morphometric measurements of prickleback fishes contributed by RC Hoover and Dr. Kassandra Ford
}
\keyword{datasets}
