% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HRData}
\alias{HRData}
\title{HRdata}
\format{
A data frame with 359 rows and 8 variables:
\describe{
\item{\code{AnimalId}}{character unique ID for each wolf}
\item{\code{PackID}}{character unique pack ID for each wolf}
\item{\code{StudyArea}}{character Study Area (BC = northeastern British Columbia, WHEC = northern Alberta, RICC = northeastern Alberta, and SK = Saskatchewan)}
\item{\code{PackSize}}{integer size of the pack}
\item{\code{Sex}}{character Sex of the individual}
\item{\code{Season}}{character Season = snow (Nov-April) or snow-free (May-October)}
\item{\code{EVIScaled}}{double primary productivity (ΔEVI, scaled between 0 and 1)}
\item{\code{Year}}{integer year of study}
\item{\code{HRsize}}{double seasonal wolf home range area (km$^2$)}
\item{\code{DiffDTScaled}}{double number of days monitored, scaled between 0 and 1}
}
}
\source{
Dickie, M., Serrouya, R., Avgar, T., McLoughlin, P., McNay, R.S., DeMars, C., Boutin, S. and Ford, A.T., Resource exploitation efficiency collapses the home range of an apex predator. Ecology, p.e3642.

Data were downloaded from https://zenodo.org/record/5643218.
}
\usage{
HRData
}
\description{
data containing the size of wolf home ranges, extent of linear features, and primary productivity (from Dickie et al. 2022)
}
\keyword{datasets}
