% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{longnosedace}
\alias{longnosedace}
\title{longnosedace}
\format{
A data frame with 68 rows and 8 variables:
\describe{
\item{\code{stream}}{character name of surveyed stream}
\item{\code{longnosedace}}{integer number of longnose dace (\emph{Rhinichthys cataractae}) in each 75-meter surveyed stream section}
\item{\code{acreage}}{integer acreage drained by the stream}
\item{\code{do2}}{double dissolved oxygen (mg/liter)}
\item{\code{maxdepth}}{integer maximum depth of the 75-meter surveyed section (cm)}
\item{\code{no3}}{double nitrate concentration (mg/liter)}
\item{\code{so4}}{double sulfate concentration (mg/liter)}
\item{\code{temp}}{double water temperature on the data the stream was surveyed (degrees C).}
}
}
\source{
Maryland Biological Stream Survey. Downloaded from \url{http://www.biostathandbook.com/multipleregression.html}
}
\usage{
longnosedace
}
\description{
Abundance data of longnose dace (\emph{Rhinichthys cataractae}) and in-stream variables collected from the Maryland Biological Stream Survey.
}
\keyword{datasets}
