% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nestocc}
\alias{nestocc}
\title{nestocc}
\format{
A data frame with 1236 rows and 7 variables:
\describe{
\item{\code{strtno}}{integer unique id for each nesting structure}
\item{\code{year}}{integer year of observation}
\item{\code{period}}{integer obsevation period: 1 = 15 March–20 April, 2= 21 April–30 April, 3= 1 May–20May, and 4= 21 May–30 June.}
\item{\code{vom}}{double visual obstruction measurements describing chronological changes in cover height and density throughout the nesting period }
\item{\code{deply}}{integer type of nest structure; deply = 1 if the structure had a single cylinder and deply = 2 if the structure had 2 cylinders}
\item{\code{nest}}{integer indicator variable equal to 1 if the structure was used during the year and observation period and 0 otherwise}
\item{\code{wetsize}}{double size of the wetland associated with the nest structure}
}
}
\source{
Zicus, M. C., Rave, D. P., Das, A., Riggs, M. R., and Buitenwerf, M. L. (2006). Influence of land use on mallard nest-structure occupancy. The Journal of wildlife management, 70(5), 1325-1333.

Zicus, M. C., D. P. Rave, and J. Fieberg. (2006). Cost effectiveness of single vs. double-cylinder over-water nest structures. Wildlife Society Bulletin 34:647-655.
}
\usage{
nestocc
}
\description{
data from a study of mallard nesting structures in Minnesota.
}
\keyword{datasets}
