% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{partialr}
\alias{partialr}
\title{partialr}
\format{
A data frame with 100 rows and 4 variables:
\describe{
\item{\code{x1}}{predictor simulated using \code{x1 = runif(100, -5, 5)}}
\item{\code{x2}}{predictor simulated using \code{x2 = 0.5*x1 + rnorm(100, 0, 2)}}
\item{\code{x3}}{predictor simulated using \code{x3= rnorm(100, 0, 3)}}
\item{\code{x4}}{predictor simulated using \code{x4 = x1*0.7 + rnorm(100, 0, 4)}}
\item{\code{y}}{response variable simulated using \code{y =  x1*1.5 - 1.2*x2 + 0.5*x3 -0.8*x3^2 + rnorm(100, 0, 5)}}
}
}
\source{
Data were simulated so that y has a linear relationship with x1, a quadratic relationship with x2, and a spurious relationship
with x3 (due to its correlation with x1).
}
\usage{
partialr
}
\description{
Simulated data set used to demonstrate added-variable and component + residual plots
}
\keyword{datasets}
