% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{peacock}
\alias{peacock}
\title{peacock}
\format{
A data frame with 131 rows and 8 variables:
\describe{
\item{\code{Place}}{character Site of observation = TZ, WZ, or CA}
\item{\code{DayofYear}}{integer Julian date}
\item{\code{Year}}{integer Year of observation}
\item{\code{Lek}}{character Lek associated with the observation}
\item{\code{MaleID}}{character Male ID}
\item{\code{TimeofHoot}}{character Time of hoot}
\item{\code{AudienceMales}}{character Indicator of whether there was  at least one other male present on the lek}
\item{\code{FemaleVisit.5m}}{integer The number of females within 5 m of each male }
}
}
\source{
Data are from Dakin, R., & Montgomerie, R. (2014). Deceptive copulation calls attract female visitors to peacock leks. The American Naturalist, 183(4), 558-564.

They were downloaded from Dryad here: https://datadryad.org/stash/dataset/doi:10.5061/dryad.vt562
}
\usage{
peacock
}
\description{
Observations of female presence before and after peacock solo hoot calls from Roslyn and Montgomerie (2013).
}
\keyword{datasets}
