% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PolikSST}
\alias{PolikSST}
\title{PolikSST}
\format{
A data frame with 139 rows and 8 variables:
\describe{
\item{\code{Site}}{integer 	Location where the core was taken during Leg 160 of the Ocean Drilling Program}
\item{\code{Sapropel}}{character Three sapropels sampled: S3 (81 ka), S5 (124 ka), and i-282 (2.943 Ma)}
\item{\code{mbsf}}{double Meters below seafloor, i.e. the depth of the sample}
\item{\code{d15N.TN}}{double The nitrogen isotopic value of the bulk sediment, in permil}
\item{\code{TEX86.SST}}{double SST reconsctructed using the TEX86 method, in degrees C}
\item{\code{Uk37.SST}}{double SST reconsctructed using the Uk37 method, in degrees C}
\item{\code{Temp.Anomaly}}{double The difference between TEX86 and Uk37 SST reconstructions
\item{\code{Sediment}}{character Whether the sample comes from inside the sapropel or from the surrounding, carbon-poor marl sediment}
}
}
\source{
These data come from the following publication:

Polik, C. A., Elling, F. J., & Pearson, A. (2018). Impacts of paleoecology on the TEX86 sea surface temperature proxy in the Pliocene‐Pleistocene Mediterranean Sea. Paleoceanography and Paleoclimatology, 33(12), 1472-1489.
}
\usage{
PolikSST
}
\description{
Sea-surface temperature (SST) data reconstructed using two different molecular signatures
}
\keyword{datasets}
