% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{predPom}
\alias{predPom}
\title{predPom}
\format{
A data frame with 99 rows and 22 variables:
\describe{
\item{\code{CaseID}}{integer  Indicates exploitation rates of a particular prey from the same study. Rows with the same CaseID represent exploitation rates by different predators from the same study.}
\item{\code{Ecosystem}}{logical  Indicates Marine or Terrestrial Ecosystem}
\item{\code{SystemID}}{integer Indicates predator-prey interactions from the same ecosystem/location and the same study. In cases where studies provided data from more than one ecosystem, different system identifiers were provided. }
\item{\code{Citation}}{character  Publication from which exploitation rates were obtained}
\item{\code{DataSource}}{character Page number and/or table from which exploitation rates were obtained within the original publication}
\item{\code{StudyType}}{character The type of study (e.g., EwE ecosystem model, survivorship, etc)  }
\item{\code{PubYear}}{integer  Year in which the study was published}
\item{\code{Region}}{character  Ocean for marine studies or continent for terrestrial studies}
\item{\code{Lat}}{double Latitude of the study}
\item{\code{NS}}{character North or South Hemisphere for Latitude}
\item{\code{Long}}{double CLongitude of the study}
\item{\code{EW}}{character East or West Hemisphere for longitude}
\item{\code{SpeciesID}}{integer nique identifier for prey species (numeric)}
\item{\code{Order}}{character  Order of prey species}
\item{\code{Family}}{character Family of prey species}
\item{\code{Genus}}{character Genus of prey species}
\item{\code{Species}}{character Prey species}
\item{\code{Paired}}{character "P" indicates studies that have exploitation estimates for both humans and non-human predators, "U" indicates exploitation estimates that are unpaired (in this file all cases are paired, "P")}
\item{\code{HumanPoM}}{double Proportion of mortality on a prey species attributed to humans}
\item{\code{MaxPOM}}{double Maximum proportion of mortality; Proportion of mortality on a prey species attributed to the highest-exploiting non-human predator in a terrestrial system}
\item{\code{CumPOM}}{double Cumulative proportion of mortality; Proportion of mortality on a prey species attributed to all non-human predators in a terrestrial system}
\item{\code{NPredPOM}}{integer Number of non-human predators in the system}
}
}
\source{
\url{https://datadryad.org/stash/dataset/doi:10.5061/dryad.238b2}

Darimont, C. T., Fox, C. H., Bryan, H. M., and Reimchen, T. E. (2015). The unique ecology of human predators. Science, 349(6250), 858-860.
}
\usage{
predPom
}
\description{
A data file with paired mortality estimates of humans and non-human predators on shared prey.
}
\keyword{datasets}
