% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RIKZdat}
\alias{RIKZdat}
\title{RIKZdat}
\format{
A data frame with 45 rows and 6 variables:
\describe{
\item{\code{week}}{integer  week of sampling (1 through 4)}
\item{\code{exposure}}{integer index of exposure determined by the surf, slope, grain size, and depth of anearobic layer}
\item{\code{NAP}}{double height of sampling station compared to mean tidal level}
\item{\code{humus}}{double fraction of organic material.}
\item{\code{Beach}}{integer unique identifier for each beach that was sampled}
\item{\code{Richness}}{integer  species richness (number of species observed)}
}
}
\source{
Zuur, A., Ieno, E.N. and Smith, G.M., 2007. Analyzing ecological data. Springer Science & Business Media.
}
\usage{
RIKZdat
}
\description{
Marine benthic data from nine inter-tidal areas along the Dutch coast. The data were collected by the Dutch institute RIKZ in the summer of 2002. In each inter-tidal area (denoted by ‘beach’), five samples were taken, and the macro-fauna and abiotic variables were measured.
}
\keyword{datasets}
