% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Selake}
\alias{Selake}
\title{Selake}
\format{
A data frame with 83 rows and 3 variables:
\describe{
\item{\code{Lake}}{character unique lake identifier}
\item{\code{Log_Water_Se}}{double concentration of Selenium in the water}
\item{\code{Log_fish_Se}}{double concentration of Selenium in a fish}
}
}
\source{
Data from Schwarz, C. J. (2014). Ch 16: Regression with pseudo-replication. In Course Notes for Beginning and Intermediate Statistics. Available at http://www.stat.sfu.ca/~cschwarz/CourseNotes. Retrieved 2015-03-1 and http://people.stat.sfu.ca/~cschwarz/Stat-650/Notes/MyPrograms/Reg-pseudo/Se-Lake/Se-lake.html
}
\usage{
Selake
}
\description{
A fictitious (but realistic) data set from Schwarz (2014) based on an environmental review of a coal mining project.
}
\keyword{datasets}
