% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{slugs}
\alias{slugs}
\title{slugs}
\format{
A data frame with 80 rows and 2 variables:
\describe{
\item{\code{slugs}}{integer count of the number of slugs under a tile}
\item{\code{field}}{character unique indicator for each grassland field (Nursery or Rookery)}
}
}
\source{
Crawley, M. J. (2012). The R book. John Wiley & Sons.\url{http://www.bio.ic.ac.uk/research/mjcraw/statcomp/data/slugsurvey.txt}
}
\usage{
slugs
}
\description{
Number of slugs found beneath 40 tiles in each of two permanent grasslands.
}
\keyword{datasets}
