% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SoilChem}
\alias{SoilChem}
\title{SoilChem}
\format{
A data frame with 69 rows and 4 variables:
\describe{
\item{\code{SampleID}}{character Sample Identifier}
\item{\code{Combined.Type}}{character Type of sediment erosional source subdivided into categories within each type to describe location of sampling or substrate material}
\item{\code{P}}{double Sediment phosphorus concentration, in milligrams per kilogram of sieved fine sediment}
\item{\code{SRP}}{double Water extractable soluble reactive phosphorus concentration, in milligrams per kilogram of sieved fine sediment}
}
}
\source{
These data are an edited subset of the data in this \href{https://conservancy.umn.edu/items/1dab3397-c604-4f4b-a9c0-93487d50d43d}{data repository} (see Barker et al. 2019).
Cases with missing values for either phosphorus measurement have been dropped, and the
varaibles have been renamed so that they can be read into R.

Combined.Types:
\itemize{
\item Ag_M = Agricultural Topsoil, Mid Slope
\item Ag_S = Agricultural Topsoil, Summit of Slope
\item Ag_To = Agricultural Topsoil, Toe of Slope
\item Bl_A = Bluff Alluvium
\item Bl_T = Bluff Till
\item CH_CH = Channel Substrate (channel bed material)
\item RV_A = Ravine Alluvium
\item RV_T = Ravine Till
\item RV_U = Ravine Upland Soil
\item SB_A = Streambank Alluvium
\item SB_T = Streambank Till
\item SB_U = Streambank Upland Soil
\item SS_L = Suspended Sediment, Lower Gage, Le Sueur River at St. Clair
\item SS_U = Suspended Sediment, Upper Gage, Le Sueur River at Minnesota highway 66
}

\strong{Reference}

Baker, Anna; Finlay, Jacques; Gran, Karen; Rorer, Michelle; Belo, Tessa; Atkins, Walter; Muramoto-Mathieu, Megumi; Yetter, Kara; Katherine, Kemmitt. (2019). Le Sueur River Basin Sediment Characterization - chemical and physical properties of sediments collected 2015-2018. Retrieved from the Data Repository for the University of Minnesota (DRUM), https://doi.org/10.13020/zbm3-9d68.
}
\usage{
SoilChem
}
\description{
Estimates of phosphorus concentration in the sediment measured in two different ways.
}
\keyword{datasets}
