% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trunkfl}
\alias{trunkfl}
\title{trunkfl}
\format{
A data frame with 619 rows and 8 variables:
\describe{
\item{\code{City}}{integer City code: 1 = Crookston, 2 = Hendricks, 3 = Hibbing, 4 = Hutchinson, 5 = Morris, 6 = Rochester}
\item{\code{Species}}{integer Species code: 2 = Acer platanoides, 3 = Acer rubrum, 4 = Acer saccharinum, 5 = Acer saccarum, 6 = Fraxinus americana, 7 = Fraxinus pennslyvanica}
\item{\code{DBH}}{integer Diameter at breast height (in)}
\item{\code{TF}}{double Trunk flare circumference (ft)}
\item{\code{BLVD}}{double Planting space width between sidewalk and street (ft)}
\item{\code{SidewalkDamage}}{integer 1 if there was sidewalk damage and 0 otherwise}
\item{\code{SGR}}{integer 100 = Observed stem girdling roots, 110 = not }
\item{\code{trunk_shape}}{character Trunk flare shape}
}
}
\source{
North, E. A., Johnson, G. R., & Burk, T. E. (2015). Trunk flare diameter predictions as an infrastructure planning tool to reduce tree and sidewalk conflicts. Urban Forestry & Urban Greening, 14(1), 65-71.
}
\usage{
trunkfl
}
\description{
data set containing measurements of trunk flare diameter and diameter at breast height for different species of trees in Minnesota cities. The data were collected for the purpose of developing predictive models
to inform management planning for city streets trees with the goal of reducing sidewalk damage.
}
\keyword{datasets}
