% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wells}
\alias{wells}
\title{wells}
\format{
A data frame with 20 rows and 6 variables:
\describe{
\item{\code{y}}{integer number of wells where TCE was detected, out of n sampled wells in total}
\item{\code{n}}{integer number of wells sampled}
\item{\code{land.use}}{character land use type associated with the well locations}
\item{\code{sewer}}{character whether or not sewers were used in the area around the wells}
\item{\code{nitrate}}{double  median concentration (mg/L) of nitrate at the N well sites}
\item{\code{chloride}}{integer median concentration (mg/L) of chloride at the N well sites}
}
}
\source{
Data were from Piegorsch and Bailer (2005) and come from a water quality study (Eckhard et al. 1989).
Eckhardt, D. A. Flipse, W. J., and Oaksford, E. T. 1989. Relation between land use and ground-quality in the upper glacial aquifer in Nassau and Suffolk Counties, Long Island. Water-Resources Investigations Report no. 86-4142. US Geological Survey, Syosset, NY.

Piegorsch, Walter W. and A. John Bailer. 2005. Analyzing Environmental Data. Wiley.
}
\usage{
wells
}
\description{
data from a water quality study of land on Long Island, New York (Eckhardt et al., 1989).
}
\keyword{datasets}
