% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WolfPackSize}
\alias{WolfPackSize}
\title{WolfPackSize}
\format{
A data frame with 660 rows and 16 variables:
\describe{
\item{\code{Year}}{integer Year of observation}
\item{\code{Pack}}{character Pack identifier}
\item{\code{Size}}{integer Pack size}
\item{\code{pack.density}}{double Density of packs per 1000 km2, w/in local vicinity (defined as 12.41 km around the territory centroid, based on the 484-km2 geometric mean territory size for wolves in MT)}
\item{\code{ungulate.summer}}{double Prey summer density index, w/in local vicinity}
\item{\code{ungulate.winter}}{double Prey winter density index, w/in local vicinity}
\item{\code{ruggedness}}{double Terrain ruggedness (vector ruggedness measure, VRM), w/in local vicinity}
\item{\code{harvest.density}}{double Harvest mortalities per 1000 km2, w/in local vicinity}
\item{\code{control.removals}}{integer Wolves removed in this pack via control removal}
\item{\code{harvest.intensity}}{character Annual harvest intensity level}
\item{\code{human.density}}{double Human density per mi$^2$, w/in local vicinity}
\item{\code{roads.lowuse}}{double Density of low-use roads, w/in local vicinity}
\item{\code{grid.removals}}{integer Wolves removed via control removal in the POM grid cell (state of MT is divided into 600km$^2$ cells, see Inman et al. 2019)}
\item{\code{mean.pack.density}}{double Mean pack density in the POM grid cell for 2005 - 2018}
\item{\code{Ecoregion}}{integer Ecoregion; 15 = Northern Rockies, 16 = Idaho Batholith, 17 = Middle Rockies, 41 = Canadian Rockies, 42 = Northwestern Glaciated Plains, 43 = Northwestern Great Plains}
\item{\code{GRID_ID}}{integer POM grid ID number}
}
}
\source{
Sells, S. N., Mitchell, M. S., Podruzny, K. M., Ausband, D. E., Emlen, D. J., Gude, J. A., ... & Loonam, K. E. (2022). Competition, prey, and mortalities influence gray wolf group size. The Journal of Wildlife Management, 86(3), e22193.

Data were originally downloaded from:  https://zenodo.org/record/5838722/export/schemaorg_jsonld
}
\usage{
WolfPackSize
}
\description{
annual pack sizes associated with wolves in Montana. Data are from Sells et al. 2022.
}
\keyword{datasets}
