% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{youtubewages}
\alias{youtubewages}
\title{youtubewages}
\format{
A data frame with 35 rows and 2 variables:
\describe{
\item{\code{VideoLink}}{character link to video of randomly selected youtuber}
\item{\code{Amount}}{character wage}
}
}
\source{
https://skewthescript.org/8-1
}
\usage{
youtubewages
}
\description{
randomly selected wages from a search of "How much do I make on YouTube" adapted from the \emph{Skew The Script curriculum} (skewthescript.org).
}
\keyword{datasets}
