% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bearmove}
\alias{bearmove}
\title{bearmove}
\format{
A data frame with 2768 rows and 10 variables:
\describe{
\item{\code{log.move}}{double log movement rate, where movement rate is calculated using the distance between successive telemetry obsevartions divided by the time between observations in meters/hour.}
\item{\code{hr}}{integer heart rate measured using implanted transmitters}
\item{\code{Season}}{character Season of the observation (Fall, Spring, or Summer)}
\item{\code{DayNight}}{character Timing of the observation (Day or Night)}
\item{\code{Sex}}{character sex of the bear (Male or Female)}
\item{\code{Stage}}{character categorical variable equal to Male, Female without cubs (Fem)), or Female with Cubs (CubPrep) Pre}
\item{\code{julian}}{integer Julian day of the observation}
\item{\code{BearID}}{integer unique identifier associated with each bear}
\item{\code{Year}}{integer Year of the observation}
\item{\code{BearIDYear}}{integer unique label assigned to each combination of bear + year}
}
}
\source{
Cardiac biologgers made by Medtronic, plc (Reveal XT Model 9529) were surgically implanted into wild American black bears (Ursus americanus) in northern Minnesota by medical professionals. The biologger data collected from these individual bears measured the heart rate ("hr") of each animal continuously but only saved one average value during each day time (collected from 08:00–20:00) and night time (collected from 00:00–04:00) period on each date ("DayNight"). The associated GPS-locations were collected using both store-on-board GPS units and Iridium satellite collars. We calculated the movement rate (meters/hour) by dividing measured step lengths (i.e., straight-line distances between consecutive GPS observations) by the length of time between locations. We only included locations collected within six hours of one another and reported the log of the mean movement rates ("log.move") occurring during each of the associated daytime and nighttime heart rate interval collection periods. See Ditmer et al. 2015 for further information on this study.

Ditmer, M.A., Garshelis, D.L., Noyce, K.V., Laske, T.G., Iaizzo, P.A., Burk, T.E., Forester, J.D. and Fieberg, J.R., 2015. Behavioral and physiological responses of American black bears to landscape features within an agricultural region. Ecosphere, 6(3), pp.1-21.
}
\usage{
bearmove
}
\description{
data set containing observed movement rates and associated average heart rates of 7 black bears in Minnesota.
}
\keyword{datasets}
