% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{birdmalariaTL}
\alias{birdmalariaTL}
\title{birdmalariaTL}
\format{
A data frame with 174 rows and 5 variables:
\describe{
\item{\code{Ind}}{character individual bird identifier}
\item{\code{Sex}}{double sex of the individual, 0 if male, 1 if female}
\item{\code{year}}{double  year of birth}
\item{\code{Age}}{double age of the bird}
\item{\code{infected}}{double infection status = 0 if uninfected, 1 if infected}
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
birdmalariaTL
}
\description{
observations of telomere length as well as demographics and infection status of great reed warblers from Asghar et al. 2015. Science 347:436-438.
}
\keyword{datasets}
