% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{package}
\name{Data4Ecologists}
\alias{Data4Ecologists}
\title{Data sets associated with Statistics for Ecologists}
\description{
This package contains data sets accompanying the book,
Statistics for ecologists.
}
